Move PageKey functions to infra/db
